+++ src/chains/Polygon.sol
@@ -5,6 +5,8 @@ import {SettlerBase} from "../SettlerBase.sol";
 import {Settler} from "../Settler.sol";
 import {SettlerMetaTxn} from "../SettlerMetaTxn.sol";
 
+import {IERC20} from "../IERC20.sol";
+import {DodoV1} from "../core/DodoV1.sol";
 import {FreeMemory} from "../utils/FreeMemory.sol";
 
 import {ISettlerActions} from "../ISettlerActions.sol";
@@ -17,6 +19,7 @@ import {
     IUniswapV3Callback
 } from "../core/univ3forks/UniswapV3.sol";
 import {IAlgebraCallback} from "../core/univ3forks/Algebra.sol";
+import {sushiswapV3PolygonFactory, sushiswapV3ForkId} from "../core/univ3forks/SushiswapV3.sol";
 import {quickSwapV3Factory, quickSwapV3InitHash, quickSwapV3ForkId} from "../core/univ3forks/QuickSwapV3.sol";
 
 // Solidity inheritance is stupid
@@ -24,7 +27,7 @@ import {SettlerAbstract} from "../SettlerAbstract.sol";
 import {AbstractContext} from "../Context.sol";
 import {Permit2PaymentAbstract} from "../core/Permit2PaymentAbstract.sol";
 
-abstract contract PolygonMixin is FreeMemory, SettlerBase {
+abstract contract PolygonMixin is FreeMemory, SettlerBase, DodoV1 {
     constructor() {
         assert(block.chainid == 137 || block.chainid == 31337);
     }
@@ -32,11 +35,21 @@ abstract contract PolygonMixin is FreeMemory, SettlerBase {
     function _dispatch(uint256 i, bytes4 action, bytes calldata data)
         internal
         virtual
-        override
+        override(SettlerAbstract, SettlerBase)
         DANGEROUS_freeMemory
         returns (bool)
     {
-        return super._dispatch(i, action, data);
+        if (super._dispatch(i, action, data)) {
+            return true;
+        } else if (action == ISettlerActions.DODOV1.selector) {
+            (IERC20 sellToken, uint256 bps, address dodo, bool quoteForBase, uint256 minBuyAmount) =
+                abi.decode(data, (IERC20, uint256, address, bool, uint256));
+
+            sellToDodoV1(sellToken, bps, dodo, quoteForBase, minBuyAmount);
+        } else {
+            return false;
+        }
+        return true;
     }
 
     function _uniV3ForkInfo(uint8 forkId)
@@ -47,6 +60,10 @@ abstract contract PolygonMixin is FreeMemory, SettlerBase {
     {
         if (forkId == uniswapV3ForkId) {
             factory = uniswapV3MainnetFactory;
+            initHash = uniswapV3InitHash;
+            callbackSelector = uint32(IUniswapV3Callback.uniswapV3SwapCallback.selector);
+        } else if (forkId == sushiswapV3ForkId) {
+            factory = sushiswapV3PolygonFactory;
             initHash = uniswapV3InitHash;
             callbackSelector = uint32(IUniswapV3Callback.uniswapV3SwapCallback.selector);
         } else if (forkId == quickSwapV3ForkId) {