+++ CHANGELOG.md
@@ -11,6 +11,12 @@ Master list of UniV3 forks:
 
 ---
 
+## 2024-06-27
+
+* Add SushiswapV3 UniV3 fork to Polygon
+* Added support for `DODOV1` action on Bnb
+* Added support for `DODOV1` action on Polygon
+
 ## 2024-06-10
 
 * Added extra anonymous argument to `execute` and `executeMetaTxn` for tracking